Introduce LikehoodBase::lnLikelihood()
We want to factor out the model evaluation from the actual computation
of the log-likelihood for forthcoming additions of "marginalized
parameters". So, we add lnValue() to LikelihoodBase, where it
calls evaluateModel() first, then we pass the model evaluation
to lnLikelihood(). Now, subclasses implement lnLikelihood with
the modelOutput as an input parameter to the function. We don't
pass modelOutput as const so that it can be overwritten to avoid
copying the modelOutput, preserving the behavior that was there
before. I'm not sure I like this behavior, but it's also a protected
function that's working a vector created internal to the class
hierarchy.