Add new evaluateModel API for likelihood marginalization
The user will need to override this method for doing marginalization
in the likelihood. In particular, this method will also pass the
current value of the marginal parameters in addition to the inference
parameters. This will be the user function that gets called if
a marginal parameter space and integration rule have been provided.